move runtime commands to subcommands
